Started work on inspector plugin
(Proof of concept)
Inspector plugin will be used to configure a node as a network object, removing need for having NetworkObject node as root, and removing need for current automatic prefab registration implementation. The plan is for NetworkObjects to be plain classes that are stored in a global dictionary with the prefab root as the key.