Working on networked nodes
When the root node of the edited scene tree is set to networked, it is registered as a networked prefab.

When a child node of the edited scene tree is set to networked, it registers itself as a networked node, and registers itself with an "owner_prefab_id", which is either the root nodes prefab id if it is a registered prefab, or -1 if not.

Prefabs which are children of the edited scene tree cannot have their networked property set, regardless of if they are registered network prefabs or not. They must have their property set by editing the actual prefab itself.

When setting the networked property of the root node of the edited scene tree, it will propagate its prefab id down to all non-prefab networked descendants, including setting the id to -1 if the property is set to false.